Refactoring: traits by command type
as_cktap on transport trait + impl Transport for Card

Fix formatting

Fix clippy warnings
